requirements.txt
flask
sqlalchemy
psycopg2-binary
requests
pydantic
python-dotenv
flask-bcrypt
pytest
